+++ src/main/java/ua/huryn/elasticsearch/view/RestaurantView.java
@@ -1,15 +1,13 @@
 package ua.huryn.elasticsearch.view;
 
-import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.router.*;
-import com.vaadin.flow.component.notification.Notification;
 import ua.huryn.elasticsearch.MainView;
 import ua.huryn.elasticsearch.service.RestaurantService;
 
 @Route(value = "restaurant", layout = MainView.class)
-public class RestaurantView extends Div implements HasUrlParameter<Integer> {
-    private int restaurantId;
+public class RestaurantView extends Div implements HasUrlParameter<Long> {
+    private Long restaurantId;
     private final RestaurantService restaurantService;
 
     public RestaurantView(RestaurantService restaurantService) {
@@ -17,7 +15,7 @@ public class RestaurantView extends Div implements HasUrlParameter<Integer> {
     }
 
     @Override
-    public void setParameter(BeforeEvent event, Integer parameter) {
+    public void setParameter(BeforeEvent event, Long parameter) {
         restaurantId = parameter;
     }
